# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Controle de Limpeza
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and numeric password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Flats ──────────────────────────────────────────────────────────────────
...
  /flats/{id}:
    get:
      operationId: getFlat
      tags: [flats]
      summary: Get a flat by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Flat details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flat"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateFlat
      tags: [flats]
      summary: Update flat occupancy status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlatUpdate"
      responses:
        "200":
          description: Updated flat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flat"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Reservations ────────────────────────────────────────────────────────────
...
  /reservations/checkouts:
    get:
      operationId: listCheckouts
      tags: [reservations]
      summary: List flats with checkouts from a given date forward
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            description: "YYYY-MM-DD date to start from (defaults to today)"
      responses:
        "200":
          description: Flats with checkouts on or after the given date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheckoutFlat"
  /reservations/checkins:
    get:
      operationId: listCheckins
      tags: [reservations]
      summary: List flats with check-ins on a given date
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            description: "YYYY-MM-DD date (defaults to today)"
      responses:
        "200":
          description: Flat IDs with check-in on given date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheckinFlat"

  # ─── Cleaning ─────────────────────────────────────────────────────────────────
...
  /cleaning/assignments/{requestId}/status:
    patch:
      operationId: updateCleaningStatus
      tags: [cleaning]
      summary: Update cleaning status for a flat (5 states)
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CleaningStatusUpdate"
      responses:
        "200":
          description: Updated cleaning assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CleaningRequest"
        "400":
          description: Observation required for pending_issue status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /cleaning/assignments/batch-claim:
    post:
      operationId: batchClaimFlats
      tags: [cleaning]
      summary: Claim multiple flats at once (will_clean status)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchClaimInput"
      responses:
        "200":
          description: Batch claim result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchClaimResult"
  /cleaning/history:
    get:
      operationId: listCleaningHistory
      tags: [cleaning]
      summary: List cleaning history with optional period filter
      parameters:
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            description: "YYYY-MM-DD start date (defaults to first day of current month)"
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            description: "YYYY-MM-DD end date (defaults to last day of current month)"
        - name: userId
          in: query
          required: false
          schema:
            type: number
            description: "Filter by user ID (admin only; camareiras see only their own)"
      responses:
        "200":
          description: Cleaning history entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CleaningHistoryEntry"

  # ─── Dashboard ───────────────────────────────────────────────────────────────
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get summary stats for a given date
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            description: "YYYY-MM-DD date (defaults to today)"
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  # ─── Settings ────────────────────────────────────────────────────────────────
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get app settings (admin only)
      responses:
        "200":
          description: App settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"

    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update app settings (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"

  # ─── Periodic Tasks ──────────────────────────────────────────────────────────
  /periodic-tasks:
    get:
      operationId: listPeriodicTasks
      tags: [periodic-tasks]
      summary: List all periodic tasks (admin only)
      responses:
        "200":
          description: List of periodic tasks with flat assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PeriodicTask"

    post:
      operationId: createPeriodicTask
      tags: [periodic-tasks]
      summary: Create a periodic task (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeriodicTaskInput"
      responses:
        "201":
          description: Created periodic task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PeriodicTask"
...
  /periodic-tasks/pending:
    get:
      operationId: listPendingPeriodicTasks
      tags: [periodic-tasks]
      summary: List pending (due or overdue) periodic tasks, optionally filtered by flat
      parameters:
        - name: flatId
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: Pending periodic tasks per flat
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PendingPeriodicTask"
  /periodic-tasks/{id}/execute:
    post:
      operationId: executePeriodicTask
      tags: [periodic-tasks]
      summary: Mark a periodic task as done for a specific flat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeriodicTaskExecutionInput"
      responses:
        "200":
          description: Execution recorded and next due date calculated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PeriodicTaskExecution"

  # ─── Observations ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: number
        username:
          type: string
        role:
          type: string
          description: "admin or camareira"

    # Flat
    Flat:
      type: object
      required: [id, number, isOccupied]
      properties:
        id:
          type: number
        number:
          type: string
          description: "Flat number/identifier"
        isOccupied:
          type: boolean
          description: "Manually controlled occupancy flag"
        updatedAt:
          type: string
    FlatUpdate:
      type: object
      properties:
        isOccupied:
          type: boolean

    # Reservations
...
    CheckoutFlat:
      type: object
      required: [flatId, flatNumber, checkoutDate, hasCheckinToday]
      properties:
        flatId:
          type: number
        flatNumber:
          type: string
        checkoutDate:
          type: string
          description: "YYYY-MM-DD"
        hasCheckinToday:
          type: boolean
          description: "Whether there's a new check-in on the checkout date (priority flag)"
        isOccupied:
          type: boolean
        cleaningRequest:
          $ref: "#/components/schemas/CleaningRequest"
    CheckinFlat:
      type: object
      required: [flatId, flatNumber, checkinDate]
      properties:
        flatId:
          type: number
        flatNumber:
          type: string
        checkinDate:
          type: string

    # Cleaning
...